fix bug

+++ bdp_batch_up_sftp.py
@@ -17,11 +17,11 @@ import zipfile
 import requests
 import threading
 import traceback
-from split_file import fsplit, new_fsplit
 from ssh_test import BdpSftpClient
 from config import Configuration
+ISOTIMEFORMAT='%Y-%m-%d %X'
 
-class FuploadHelper(object):
+class FuploadHelper2(object):
     def __init__(self, domain, username, password):
         self.domain = domain
         self.username = username
@@ -30,7 +30,7 @@ class FuploadHelper(object):
         self.user_id = self.login()
         self.flag = False
         self.total = 0
-        #self.total = 1
+        self.chunksize = int(self.conf.get('server', 'chunksize'))* 1024*1024
 
     def login(self):
         data = dict(
@@ -51,7 +51,7 @@ class FuploadHelper(object):
         else:
             return result['result']['user_id']
 
-    def do_action(self, local_path):
+    def create(self, local_path):
 
         # 第一步 获取schema信息 创建数据源
         ds_name, tbs = self.get_dsinfo(local_path)
@@ -65,20 +65,37 @@ class FuploadHelper(object):
         ezio_create = self.conf.get('server', 'Ezio') + '/api/batchcreate'
         res = requests.post(ezio_create, bag)
         print res.content
+	print 
         result = json.loads(res.content)['result']
-        user_id = result['user_id']
         ds_id = result['ds_id']
         db_id = result['db_id']
         tb_res = result['tbs']
+        sconf = self.conf.get_section('sftp')
+        bsftp = BdpSftpClient(sconf['host'],int(sconf['port']),sconf['username'],sconf['password'])
 
-        # 第三步 将路径下的数据全部分割 
+        # 第三步 将路径下的数据全部分割
         for tbinfo in tb_res:
-            tb_id = tbinfo['tb_id']
-            schema = tbinfo['schema']
-            separator = tbinfo['separator']
-            null_holder = tbinfo['null_holder']
-            chunksize = 5 * 1024 * 1024
-            sub_path = os.path.join(local_path, tbinfo['tbname'])
+            path = os.path.join(local_path, tbinfo['tbname'])
+            remotepath = '/tmp/%s/%s' % (ds_id, tbinfo['tb_id'])
+	    bsftp.ssh_exec_command('mkdir -p %s' % remotepath)
+            print path
+            self.do_action(ds_id, tbinfo['tb_id'], tbinfo['schema'], tbinfo['separator'], tbinfo['null_holder'], path)
+
+    def append(self, local_path, ds_name, tbname):
+        bag = dict(
+                user_id=self.user_id,
+                ds_name=ds_name,
+                tbname=tbname)
+        ezio_info = self.conf.get('server', 'Ezio') + '/new/info'
+        res = requests.post(ezio_info, bag)
+        print res.content
+        result = json.loads(res.content)['result']
+        ds_id = result['ds_id']
+        tb_id = result['tb_id']
+        separator, null_holder, schema = self.get_schema(local_path)
+        self.do_action(ds_id, tb_id, schema, separator, null_holder, local_path)
+
+    def do_action(self, ds_id, tb_id, schema, separator, null_holder,sub_path, err=None):
             thread = []
             self.flag = False
             queue = Queue.Queue()
@@ -86,7 +103,7 @@ class FuploadHelper(object):
 
             # 边读取边发送
             # queue (num, total, data)
-            th0 = threading.Thread(target=self.split_file, args=(queue, sub_path, chunksize))
+            th0 = threading.Thread(target=self.split_file, args=(queue, sub_path, err))
             thread.append(th0)
 
             # 第四步 多线程实现文件上传
@@ -103,7 +120,7 @@ class FuploadHelper(object):
             cres =  self.check(ds_id, tb_id)
             if cres:
                 error_bag = {
-                        'user_id':user_id,
+                        'user_id':self.user_id,
                         'ds_id':ds_id,
                         'tb_id':tb_id,
                         'err':cres,
@@ -114,15 +131,15 @@ class FuploadHelper(object):
                 with open('%s/err.log' % sub_path , 'wb') as  fo:
                     fo.write('%s' % json.dumps(error_bag))
                 print '文件发送失败:%s' % json.dumps(cres)
-                continue
             else:
-                task_id = self.merge(user_id, ds_id, tb_id, schema, separator, null_holder)
+                task_id = self.merge(ds_id, tb_id, schema, separator, null_holder)
                 print '生成任务task_id:%s' % str(task_id)
-        return True
+                self.delete(sub_path)
+            return True
 
-    def merge(self, user_id, ds_id, tb_id, schema, separator, null_holder):
+    def merge(self, ds_id, tb_id, schema, separator, null_holder):
         bag = dict(
-                user_id=user_id,
+                user_id=self.user_id,
                 ds_id=ds_id,
                 tb_id=tb_id,
                 total=self.total,
@@ -143,51 +160,62 @@ class FuploadHelper(object):
                 tb_id=tb_id,
                 total=self.total
                 )
-        ezio_check = self.conf.get('server', 'Ezio') + '/new/check'
+        ezio_check = self.conf.get('server', 'Ezio') + '/new/newcheck'
         res = requests.post(ezio_check, bag)
         res = json.loads(res.content)['result']
         return res
 
-    def split_file(self, queue, sub_path, chunksize):
-        res = os.listdir(sub_path)
+    def split_file(self, queue, sub_path, err=None):
+        print sub_path
         zipname = os.path.join(sub_path, 'all.zip')
-        # 过滤压缩无关文件
-        if "schema.info" in res:
-            res.remove('schema.info')
-        if "err.log" in res:
-            res.remove('err.log')
-        if "all.zip" in res:
-            res.remove('all.zip')
-            os.remove(zipname)
-        # 压缩该文件下所有需要上传的文件
-        f = zipfile.ZipFile(zipname,'w',zipfile.ZIP_DEFLATED)
-        for fname in res:
-            fromfile = os.path.join(sub_path, fname)
-            if os.path.isdir(fromfile):
-                continue
-            f.write(fromfile)
-        f.close()
-
+        print zipname
+        if err is None:
+            res = os.listdir(sub_path)
+            # 过滤压缩无关文件
+            if "schema.info" in res:
+                res.remove('schema.info')
+            if "err.log" in res:
+                res.remove('err.log')
+            if "all.zip" in res:
+                res.remove('all.zip')
+                os.remove(zipname)
+            # 压缩该文件下所有需要上传的文件
+            f = zipfile.ZipFile(zipname,'w',zipfile.ZIP_DEFLATED)
+            for fname in res:
+                fromfile = os.path.join(sub_path, fname)
+                if os.path.isdir(fromfile):
+                    continue
+                f.write(fromfile)
+            f.close()
         size = os.path.getsize(zipname)
-        total = int(math.ceil(float(size)/chunksize))
+        total = int(math.ceil(float(size)/self.chunksize))
         self.total = total
-        md5sum = os.popen('md5 %s' % zipname).read()
-        md5 = md5sum.split('= ')[1].split('\n')[0]
+        #md5sum = os.popen('md5 %s' % zipname).read()
+        #md5 = md5sum.split('= ')[1].split('\n')[0]
         partnum = 0
+        if err is None:
+            err = [i+1  for i in range(total)]
         input = open(zipname, 'rb')
         while 1:
-            chunk = input.read(chunksize)
+            chunk = input.read(self.chunksize)
             if not chunk:
                 break
             partnum += 1
-            bag = {}
-            bag['fname'] = 'all.zip'
-            bag['partnum'] = partnum
-            bag['total'] = total
-            bag['size'] = size
-            bag['das'] = chunk
-            bag['md5'] = md5
-            queue.put(bag)
+            filename = os.path.join(sub_path, ('all.zip-%s' % partnum))
+            fileobj = open(filename, 'wb')
+            fileobj.write(chunk)
+            fileobj.close()
+
+            if partnum in err:
+                bag = {}
+                bag['fname'] = "all.zip"
+                bag['partnum'] = partnum
+                bag['total'] = total
+                bag['size'] = size
+                bag['das'] = filename
+                #bag['md5'] = md5
+                queue.put(bag)
+        input.close()
 
         while True:
             if queue.qsize() == 0:
@@ -198,7 +226,7 @@ class FuploadHelper(object):
         return True
 
     def send_file(self, queue, remotepath):
-        ezio_upload  = Configuration().get('server', 'Ezio') + '/new/upload'
+        #ezio_upload  = Configuration().get('server', 'Ezio') + '/new/upload'
         while True:
             info = None
             try:
@@ -210,7 +238,12 @@ class FuploadHelper(object):
                     time.sleep(1)
             if info:
                 info['path'] = remotepath
-                res = requests.post(ezio_upload, data=info)
+                sconf = self.conf.get_section('sftp')
+                bsftp = BdpSftpClient(sconf['host'],int(sconf['port']),sconf['username'],sconf['password'])
+                rpath = os.path.join(remotepath, 'all.zip-%s' % info['partnum'])
+                bsftp.sftp_put(info['das'], rpath)
+                bsftp.close()
+                #res = requests.post(ezio_upload, data=info)
         return True
 
     def get_schema(self, local_path):
@@ -235,9 +268,11 @@ class FuploadHelper(object):
         return separator, null_holder, res
 
     def delete(self, path):
+        print path
         res = os.listdir(path)
         for item in res:
-            os.remove(item)
+            fpname = os.path.join(path, item)
+            os.remove(fpname)
 
     def get_dsinfo(self, path):
         res = os.listdir(path)
@@ -282,90 +317,20 @@ class FuploadHelper(object):
                 temp = line.split('\n')[0]
                 bag = json.loads(temp)
                 break
-            user_id = bag['user_id']
-            ds_id = bag['ds_id']
-            tb_id = bag['tb_id']
-            err = bag['err']
-            schema = bag['schema']
-            separator = bag['separator']
-            null_holder = bag['null_holder']
-            chunksize = 5*1024*1024
-            thread = []
-            self.flag = False
-            queue = Queue.Queue()
-            remotepath = '/tmp/%s/%s' % (ds_id, tb_id)
-            
-            th0 = threading.Thread(target=self.retry_split_file, args=(queue, sub_path, chunksize, err))
-            thread.append(th0)
-            for i in range(5):
-                th = threading.Thread(target=self.send_file, args=(queue, remotepath))
-                thread.append(th)
-            for t in thread:
-                t.setDaemon(True)
-                t.start()
-            for t in thread:
-                t.join()
-            print "发送完毕"
-            cres =  self.check(ds_id, tb_id)
-            if cres:
-                error_bag = {
-                        'user_id':user_id,
-                        'ds_id':ds_id,
-                        'tb_id':tb_id,
-                        'err':cres,
-                        'schema':schema,
-                        'separator':separator,
-                        'null_holder':null_holder
-                        }
-                with open('%s/err.log' % sub_path , 'wb') as  fo:
-                    fo.write('%s' % json.dumps(error_bag))
-                print '文件发送失败:%s' % json.dumps(cres)
-                continue
-            else:
-                task_id = self.merge(user_id, ds_id, tb_id, schema, separator, null_holder)
-                print '生成任务task_id:%s' % str(task_id)
+            self.do_action(bag['ds_id'], bag['tb_id'], bag['schema'], bag['separator'], bag['null_holder'], sub_path, bag['err'])
         return True
-
-    def retry_split_file(self, queue, sub_path, chunksize, err):
-            zipname = os.path.join(sub_path, 'all.zip')
-            size = os.path.getsize(zipname)
-            total = int(math.ceil(float(size)/chunksize))
-            self.total = total
-            md5sum = os.popen('md5 %s' % zipname).read()
-            md5 = md5sum.split('= ')[1].split('\n')[0]
-            partnum = 0
-            input = open(zipname, 'rb')
-            while 1:
-                chunk = input.read(chunksize)
-                if not chunk:
-                    break
-                partnum += 1
-                if partnum in err:
-                    bag = {}
-                    bag['fname'] = 'all.zip'
-                    bag['partnum'] = partnum
-                    bag['total'] = total
-                    bag['size'] = size
-                    bag['das'] = chunk
-                    bag['md5'] = md5
-                    queue.put(bag)
-            while True:
-                if queue.qsize() == 0:
-                    self.flag = True
-                    break
-                else:
-                    time.sleep(5)
-            return True
 
 if __name__=='__main__':
     domain = "haizhi"
     username = "jiaoguofu"
     password = "jiao1993"
-    local_path = "/Users/jiaoguofu/Desktop/fsplit"
+    local_path = "/Users/jiaoguofu/Desktop/fsplit_bak/tb"
+    ds_name = "fsplit_bak"
+    tbname = "tb"
     import datetime
     print datetime.datetime.today()
-    #FuploadHelper(domain, username, password).do_action(local_path)
-    FuploadHelper(domain, username, password).retry(local_path)
+    #FuploadHelper(domain, username, password).create(local_path)
+    FuploadHelper(domain, username, password).append(local_path, ds_name, tbname)
+    #FuploadHelper(domain, username, password).retry(local_path)
     print datetime.datetime.today()
-
